Move `changedAttributes` tests to their own file
